[refactor] :recycle: realocate public functions used in one place (#276)
