+++ src/services/predictWalletService.ts
@@ -5,7 +5,7 @@ import { Logger } from '../helpers/loggerHelper';
 import { SIGNING_KEY } from '../config/constants';
 import { IBlockchain } from '../models/blockchain';
 import { getNetworkConfig } from './networkService';
-import { getDynamicGas } from '../helpers/dynamicGasHelper';
+import { getDynamicGas } from '../helpers/paymasterHelper';
 import { getChatterPayWalletFactoryABI } from './bucketService';
 import { generatePrivateKey } from '../helpers/keyGeneratorHelper';
 import { ChatterPayWalletFactory__factory } from '../types/ethers-contracts';
